Role: minor fix
