+++ lib/frontend/src/actions/index.js
@@ -21,20 +21,6 @@ export const setParams = params => ({
   params,
 })
 
-export function editClickItem(itemId) {
-  return {
-    type: 'EDIT_ITEM',
-    itemId: itemId,
-  }
-}
-
-export function cancelClickItem(itemId) {
-  return {
-    type: 'CANCEL_ITEM',
-    itemId: itemId,
-  }
-}
-
 export const checkAccountabilities = content => ({
   type: 'CHECK_ACCOUNTABILITIES',
   content,
@@ -42,18 +28,6 @@ export const checkAccountabilities = content => ({
 
 export const delAccountabilities = () => ({
   type: 'DEL_VALUE',
-})
-
-export const editRole = () => ({
-  type: 'EDIT_ROLE',
-})
-
-export const checkForm = () => ({
-  type: 'CHECKLIST_ITEM',
-})
-
-export const indicatorForm = () => ({
-  type: 'INDICATOR_ITEM',
 })
 
 export const setUser = user => ({
