+++ lib/frontend/src/actions/roles.js
@@ -1,6 +1,32 @@
 import { roleNameFromState } from '../utils'
 import { setError, setResults, fetchResults } from './index'
 
+export const editRole = () => ({
+  type: 'EDIT_ROLE',
+})
+
+export function editClickItem(itemId) {
+  return {
+    type: 'EDIT_ITEM',
+    itemId: itemId,
+  }
+}
+
+export function cancelClickItem(itemId) {
+  return {
+    type: 'CANCEL_ITEM',
+    itemId: itemId,
+  }
+}
+
+export const checkForm = () => ({
+  type: 'CHECKLIST_ITEM',
+})
+
+export const indicatorForm = () => ({
+  type: 'INDICATOR_ITEM',
+})
+
 export const createRole = (data, history) => async dispatch => {
   let response
   try {
@@ -52,8 +78,9 @@ export const deleteRole = (roleId, circleId, history) => async dispatch => {
       },
     })
   } catch (e) {
+    console.error(e)
     return dispatch(
-      setError(`Erreur pendant la récupération des données, "${roleId}"`)
+      setError(`Erreur pendant la suppression du role "${roleId}"`, 'role')
     )
   }
   if (response.status >= 200 && response.status < 300) {
@@ -61,6 +88,7 @@ export const deleteRole = (roleId, circleId, history) => async dispatch => {
       await response.json()
       return history.push(`/circle?circle_id=${circleId}`)
     } catch (e) {
+      console.error(e)
       return dispatch(
         setError(
           `La réponse ne contient pas de json pour le cercle "${
